Code refactoring. Split up AO.js into three different material files
